requirements.txt
joblib
matplotlib
numpy
openpyxl
pandas
psycopg2-binary
python-dotenv
scikit-learn
seaborn
tensorflow
torch
boto3

fastapi 
uvicorn 
sqlalchemy 
asyncpg